+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.7.0 (IN PROGRESS)
 * Add action options to choose jobs screen (UIDATIMP-268)
+* Add non-editable mode feature for RecordTypesSelect component (UIDATIMP-323)
 
 ## [1.6.0](https://github.com/folio-org/ui-data-import/tree/v1.6.0) (2019-11-01)
 * Add details section on create/edit action profile form (UIDATIMP-207)